+++ wasm/codegen/generators/constants.py
@@ -660,7 +660,7 @@ FUNCTION_BOUNDS_CHECKS: Dict[str, str] = {
   CHECK_SIZE(DsDq, m.nv() * m.nsensordata());
   CHECK_SIZE(DsDv, m.nv() * m.nsensordata());
   CHECK_SIZE(DsDa, m.nv() * m.nsensordata());
-  CHECK_SIZE(DmDq, m.nv() * m.nM());
+  CHECK_SIZE(DmDq, m.nv() * m.nC());
     """.strip(),
     "mjd_subQuat": """
   CHECK_SIZE(qa, 4);